+++ projects/igniteui-angular/src/lib/combo/combo.component.spec.ts
@@ -1058,6 +1058,7 @@ describe('igxCombo', () => {
                 expect(input.nativeElement.getAttribute('aria-expanded')).toMatch('false');
                 expect(input.nativeElement.getAttribute('aria-controls')).toEqual(combo.dropdown.listId);
                 expect(input.nativeElement.getAttribute('aria-labelledby')).toEqual(combo.placeholder);
+                expect(input.nativeElement.getAttribute('aria-label')).toEqual('No options selected');
 
                 const dropdown = fixture.debugElement.query(By.css(`div[role="listbox"]`));
                 expect(dropdown.nativeElement.getAttribute('aria-labelledby')).toEqual(combo.placeholder);
@@ -1073,12 +1074,16 @@ describe('igxCombo', () => {
 
                 const list = fixture.debugElement.query(By.css(`.${CSS_CLASS_CONTENT}`));
                 expect(list.nativeElement.getAttribute('aria-multiselectable')).toEqual('true');
-                expect(list.nativeElement.getAttribute('aria-activedescendant')).toEqual('');
+                expect(list.nativeElement.getAttribute('aria-activedescendant')).toEqual(null);
 
                 UIInteractions.triggerEventHandlerKeyDown('ArrowDown', list);
                 tick();
                 fixture.detectChanges();
                 expect(list.nativeElement.getAttribute('aria-activedescendant')).toEqual(combo.dropdown.focusedItem.id);
+
+                combo.select(['Illinois', 'Mississippi', 'Ohio']);
+                fixture.detectChanges();
+                expect(input.nativeElement.getAttribute('aria-label')).toEqual('Selected options');
             }));
             it('should render aria-expanded attribute properly', fakeAsync(() => {
                 expect(input.nativeElement.getAttribute('aria-expanded')).toMatch('false');
@@ -1938,6 +1943,16 @@ describe('igxCombo', () => {
                     fixture.detectChanges();
                     expect(firstVisibleItem.classList.contains(CSS_CLASS_FOCUSED)).toBeTruthy();
                 }));
+                it('should close the dropdown list on pressing Tab key', fakeAsync(() => {
+                    combo.toggle();
+                    fixture.detectChanges();
+
+                    const dropdownContent = fixture.debugElement.query(By.css(`.${CSS_CLASS_CONTENT}`));
+                    UIInteractions.triggerEventHandlerKeyDown('Tab', dropdownContent);
+                    tick();
+                    fixture.detectChanges();
+                    expect(combo.collapsed).toBeTruthy();
+                }));
             });
             describe('primitive data dropdown: ', () => {
                 it('should properly navigate with HOME/END keys when no virtScroll is necessary', async () => {
@@ -2268,6 +2283,37 @@ describe('igxCombo', () => {
                         displayText: selectedItem_2.value[combo.valueKey],
                         cancel: false
                     });
+            });
+            it('should toggle combo dropdown on Enter of the focused toggle icon', fakeAsync(() => {
+                spyOn(combo, 'toggle').and.callThrough();
+                const toggleBtn = fixture.debugElement.query(By.css(`.${CSS_CLASS_TOGGLEBUTTON}`));
+
+                UIInteractions.triggerEventHandlerKeyDown('Enter', toggleBtn);
+                tick();
+                fixture.detectChanges();
+                expect(combo.toggle).toHaveBeenCalledTimes(1);
+                expect(combo.collapsed).toEqual(false);
+
+                UIInteractions.triggerEventHandlerKeyDown('Enter', toggleBtn);
+                tick();
+                fixture.detectChanges();
+                expect(combo.toggle).toHaveBeenCalledTimes(2);
+                expect(combo.collapsed).toEqual(true);
+            }));
+            it('should clear the selection on Enter of the focused clear icon', () => {
+                const selectedItem_1 = combo.dropdown.items[1];
+                combo.toggle();
+                fixture.detectChanges();
+                simulateComboItemClick(1);
+                expect(combo.selection[0]).toEqual(selectedItem_1.value);
+                expect(combo.value[0]).toEqual(selectedItem_1.value[combo.valueKey]);
+
+                const clearBtn = fixture.debugElement.query(By.css(`.${CSS_CLASS_CLEARBUTTON}`));
+                UIInteractions.triggerEventHandlerKeyDown('Enter', clearBtn);
+                fixture.detectChanges();
+                expect(input.nativeElement.value).toEqual('');
+                expect(combo.selection.length).toEqual(0);
+                expect(combo.value.length).toEqual(0);
             });
             it('should not be able to select group header', () => {
                 spyOn(combo.selectionChanging, 'emit').and.callThrough();