+++ CHANGELOG.md
@@ -29,6 +29,15 @@ All notable changes for each version of this project will be documented in this 
 
     It's possible also to set a `mergeComparer` on the individual columns, in case some custom handling is needed for a particular data field.
 
+- `IgxCarousel`
+    - Added `select` method overload accepting index.
+    ```ts
+    this.carousel.select(2, Direction.NEXT);
+    ```
+
+### General
+- `IgxDropDown` now exposes a `role` input property, allowing users to customize the role attribute based on the use case. The default is `listbox`.
+
 
 ## 20.0.6
 ### General