+++ CHANGELOG.md
@@ -3,8 +3,16 @@
 All notable changes for each version of this project will be documented in this file.
 
 ## 20.1.0
+### New Features
+- `IgxCarousel`
+    - Added `select` method overload accepting index.
+    ```ts
+    this.carousel.select(2, Direction.NEXT);
+    ```
+
 ### General
 - `IgxDropDown` now exposes a `role` input property, allowing users to customize the role attribute based on the use case. The default is `listbox`.
+
 
 ## 20.0.6
 ### General